ref(ui): move remains from deprecatedSmartSearchBar into components/searchBar
as that's where they are mostly used
