+++ static/app/components/searchBar/index.tsx
