+++ static/app/components/searchBar/types.tsx
