+++ static/app/components/searchBar/utils.tsx
